@@ -3,7 +3,7 @@ apply from: 'global.gradle'
 
 buildscript {
     ext.objectboxVersion = '2.5.1'
-    ext.kotlin_version = "1.3.72"
+    ext.kotlin_version = "1.5.21"
     repositories {
         google()
         mavenCentral()
