@@ -8,7 +8,7 @@ import androidx.appcompat.widget.Toolbar
 import com.android.exercise.R
 import com.android.exercise.base.BaseActivity
 import com.android.exercise.base.toolbar.ToolBarCommonHolder
-import kotlinx.android.synthetic.main.activity_bitmap_optimize.*
+import com.android.exercise.databinding.ActivityBitmapOptimizeBinding
 
 /**
  * 大图加载
@@ -17,10 +17,12 @@ import kotlinx.android.synthetic.main.activity_bitmap_optimize.*
  * Created by wangzhen on 2020/7/8.
  */
 class BitmapOptimizeActivity : BaseActivity() {
+    private lateinit var binding: ActivityBitmapOptimizeBinding
     private var mBitmap: Bitmap? = null
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_bitmap_optimize)
+        binding = ActivityBitmapOptimizeBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
         raw()
         inSampleSize()
@@ -33,22 +35,23 @@ class BitmapOptimizeActivity : BaseActivity() {
         options.inMutable = true
         options.inSampleSize = 1
         val bitmap = BitmapFactory.decodeResource(resources, R.mipmap.bicycle, options)
-        iv_in_bitmap.setImageBitmap(bitmap)
-        tv_in_bitmap.text = String.format("内存重用(%s)", formatSize(bitmap))
+        binding.ivInBitmap.setImageBitmap(bitmap)
+        binding.tvInBitmap.text = String.format("内存重用(%s)", formatSize(bitmap))
     }
 
     private fun inSampleSize() {
         val options = BitmapFactory.Options()
         options.inSampleSize = 2
         val bitmap = BitmapFactory.decodeResource(resources, R.mipmap.bicycle, options)
-        iv_in_sample_size.setImageBitmap(bitmap)
-        tv_in_sample_size.text = String.format("修改采样率(%s)\ninSampleSize = 2\n大小缩小1/2\n质量缩小1/4", formatSize(bitmap))
+        binding.ivInSampleSize.setImageBitmap(bitmap)
+        binding.tvInSampleSize.text =
+            String.format("修改采样率(%s)\ninSampleSize = 2\n大小缩小1/2\n质量缩小1/4", formatSize(bitmap))
     }
 
     private fun raw() {
         mBitmap = BitmapFactory.decodeResource(resources, R.mipmap.bicycle)
-        iv_raw.setImageBitmap(mBitmap)
-        tv_raw.text = String.format("原图(%s)", formatSize(mBitmap))
+        binding.ivRaw.setImageBitmap(mBitmap)
+        binding.tvRaw.text = String.format("原图(%s)", formatSize(mBitmap))
     }
 
     private fun formatSize(bitmap: Bitmap?): String {
