@@ -12,7 +12,7 @@ import com.android.exercise.R
 import com.android.exercise.base.BaseActivity
 import com.android.exercise.base.okhttp.OKHttpManager
 import com.android.exercise.base.toolbar.ToolBarCommonHolder
-import kotlinx.android.synthetic.main.activity_web_socket.*
+import com.android.exercise.databinding.ActivityWebSocketBinding
 import okhttp3.Request
 import okhttp3.Response
 import okhttp3.WebSocket
@@ -26,6 +26,7 @@ import java.util.concurrent.Executors
  * Created by wangzhen on 2020/8/15.
  */
 class WebSocketActivity : BaseActivity() {
+    private lateinit var binding: ActivityWebSocketBinding
     private val mWebServer = MockWebServer()
     private var mClientSocket: WebSocket? = null
     private var mServerSocket: WebSocket? = null
@@ -45,9 +46,10 @@ class WebSocketActivity : BaseActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_web_socket)
+        binding = ActivityWebSocketBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
-        tv_msg.text = "-------OkHttp 长连接-------\n"
+        binding.tvMsg.text = "-------OkHttp 长连接-------\n"
         Executors.newCachedThreadPool().submit {
             initWebServer()
             initClient()
@@ -58,32 +60,33 @@ class WebSocketActivity : BaseActivity() {
     private fun initClient() {
         val url = "ws://" + mWebServer.hostName + ":" + mWebServer.port
         Log.e(TAG, "initClient: $url")
-        OKHttpManager.getClient().newWebSocket(Request.Builder().url(url).build(), object : WebSocketListener() {
-            override fun onOpen(webSocket: WebSocket, response: Response) {
-                super.onOpen(webSocket, response)
-                mClientSocket = webSocket
-                mHandler.post {
-                    tv_msg.append("客户端：连接成功\n")
+        OKHttpManager.getClient()
+            .newWebSocket(Request.Builder().url(url).build(), object : WebSocketListener() {
+                override fun onOpen(webSocket: WebSocket, response: Response) {
+                    super.onOpen(webSocket, response)
+                    mClientSocket = webSocket
+                    mHandler.post {
+                        binding.tvMsg.append("客户端：连接成功\n")
+                    }
+                    webSocket.send("我是客户端，你好呀")
                 }
-                webSocket.send("我是客户端，你好呀")
-            }
 
-            override fun onMessage(webSocket: WebSocket, text: String) {
-                super.onMessage(webSocket, text)
-                mHandler.post {
-                    tv_msg.append("客户端：$text\n")
-                    scroller.fullScroll(ScrollView.FOCUS_DOWN)
+                override fun onMessage(webSocket: WebSocket, text: String) {
+                    super.onMessage(webSocket, text)
+                    mHandler.post {
+                        binding.tvMsg.append("客户端：$text\n")
+                        binding.scroller.fullScroll(ScrollView.FOCUS_DOWN)
+                    }
                 }
-            }
 
-            override fun onClosed(webSocket: WebSocket, code: Int, reason: String) {
-                super.onClosed(webSocket, code, reason)
-                mClientSocket = null
-                mHandler.post {
-                    tv_msg.append("客户端：断开连接\n")
+                override fun onClosed(webSocket: WebSocket, code: Int, reason: String) {
+                    super.onClosed(webSocket, code, reason)
+                    mClientSocket = null
+                    mHandler.post {
+                        binding.tvMsg.append("客户端：断开连接\n")
+                    }
                 }
-            }
-        })
+            })
     }
 
     private fun initWebServer() {
@@ -92,7 +95,7 @@ class WebSocketActivity : BaseActivity() {
                 super.onOpen(webSocket, response)
                 mServerSocket = webSocket
                 mHandler.post {
-                    tv_msg.append("服务端：连接成功\n")
+                    binding.tvMsg.append("服务端：连接成功\n")
                 }
                 webSocket.send("我是服务器，你好呀")
             }
@@ -100,7 +103,7 @@ class WebSocketActivity : BaseActivity() {
             override fun onMessage(webSocket: WebSocket, text: String) {
                 super.onMessage(webSocket, text)
                 mHandler.post {
-                    tv_msg.append("服务端：$text\n")
+                    binding.tvMsg.append("服务端：$text\n")
                 }
             }
 
@@ -108,7 +111,7 @@ class WebSocketActivity : BaseActivity() {
                 super.onClosed(webSocket, code, reason)
                 mServerSocket = null
                 mHandler.post {
-                    tv_msg.append("服务端：断开连接\n")
+                    binding.tvMsg.append("服务端：断开连接\n")
                 }
 
             }
