@@ -10,24 +10,31 @@ import androidx.recyclerview.widget.RecyclerView
 import com.android.exercise.R
 import com.android.exercise.base.BaseActivity
 import com.android.exercise.base.toolbar.ToolBarCommonHolder
+import com.android.exercise.databinding.ActivitySharedElementsBinding
 import com.android.exercise.ui.activity.shared_elements.adapter.GridAdapter
 import com.android.exercise.ui.activity.shared_elements.entity.GridEntity
-import kotlinx.android.synthetic.main.activity_shared_elements.*
 
 /**
  * 共享元素
  * Created by wangzhen on 2020/7/20.
  */
 class SharedElementsActivity : BaseActivity() {
+    private lateinit var binding: ActivitySharedElementsBinding
     private val _spanCount: Int = 3
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_shared_elements)
+        binding = ActivitySharedElementsBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
-        recycler.layoutManager = GridLayoutManager(this, _spanCount)
-        recycler.addItemDecoration(object : RecyclerView.ItemDecoration() {
-            override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
+        binding.recycler.layoutManager = GridLayoutManager(this, _spanCount)
+        binding.recycler.addItemDecoration(object : RecyclerView.ItemDecoration() {
+            override fun getItemOffsets(
+                outRect: Rect,
+                view: View,
+                parent: RecyclerView,
+                state: RecyclerView.State
+            ) {
                 super.getItemOffsets(outRect, view, parent, state)
                 val divider = view.resources.getDimensionPixelOffset(R.dimen.divider_size)
 
@@ -37,7 +44,7 @@ class SharedElementsActivity : BaseActivity() {
                 outRect.bottom = divider
             }
         })
-        recycler.adapter = GridAdapter(list())
+        binding.recycler.adapter = GridAdapter(list())
     }
 
     private fun list(): MutableList<GridEntity> {
